Proper treatment of realized functions.
Previously, KreMLin would choke on functions that advertised a non-void return
type, yet had an empty body. Such functions would arise from "assume val ...".

Now, the body of these functions is a single "Abort" node, which extracts to
"exit(255)", a much cleaner way of dealing with the problem (rather than ignore
it and extract the function with an empty body).

From cursory testing, it seems like "exit" is marked as "__noreturn" and does
not generate a warning.