Applying an offset to a location is now part of the syntax
"loc (b, (n + n') + n'', [])" is now represented as "loc (b, n, [n'', n'], [])".

This seems to be required to prove the initialization lemma for the relation R,
stating that a lowstar expression is related to its translation.

In cases readbuf and writebuf, in lowstar, the base pointer and the offset are
two separate arguments of the function. In Cstar, they get evaluated to a single
location. Then, the back-translation needs to be able to take apart the base
pointer and offset from this location, in order to back-translate to a lowstar
expression.