A very nasty bug.
Consider the following, legit assignment:

x <- bufcreate 1 e[x]

where e[x] denotes an expression that references x.

The scopes are correct, however, the right-hand side of this assignment,
might be evaluated too late, because, this gets compiled as:

t *x = malloc(sizeof t);
x[0] = e[x]

meaning in the second line x evaluates to the new address as opposed to
the old value. To mitigate this, the set of variables present in the
right-hand side is computed, and if x is present, the initial expression
is forcefully hoisted to guarantee proper evaluation.