Strengthen the invariants for the Wasm compilation of struct literals.
In C11, a compound literal can appear at any position in an expression.
An occurrence of { f = e; ... } is syntactic sugar for:

  struct t x = { f = e; ... };

  ... x ...

where the allocation occurs in the nearest enclosing scope. This,
itself, is syntactic sugar for an uninitialized stack allocation
followed by writing the initial value at the allocated address.

The Wasm backend now performs this desugaring in two steps.
- in Structs.ml, to_addr materializes the first step and establishes
  that:
  * the only occurrences of struct literals are underneath EBufCreate
    nodes
  * these struct literals may themselves contain nested struct literals.
- in SimplifyWasm.ml, we perform the second step of desugaring and
  ensure that EBufCreate is desugared into EBufCreate EAny + EBufWrite.

The first step is crucial (see test/Wasm6.fst); leaving occurrences of
struct literals anywhere else (e.g. in EBufWrite nodes) would violate
the evaluation semantics and result in miscompilations.

The hoisting phase in Simplify.ml now superbly ignores struct literals
(ECons nodes) and leaves it up to the Wasm backend to perform proper
hoisting. This is a better separation of concerns.