Fix a bug related to short-circuiting boolean operators.
In e1 && e2, or e1 || e2, if e2, once translated to C*, gives rise to
let-bindings, KReMLin would previously error out and request a rewrite
of the offending code.

It turns out that this error was not caught when the boolean expression
was in terminal position. This patches fixes this.

This in turn generates a handful of errors in miTLS. Rather than rewrite
everything, KReMLin now applies the following rewrite rules on the fly:
  e1 && e2 --> if e1 then e2 else false
  e1 || e2 --> if e1 then true else e2
whenever e2 generates let-bindings.