Add special treatment for Lib.Memzero0.memzero
The function is an assume val in F*, and these do not get extracted to
KreMLin. So, the first thing we do is enrich the initial environment
with:

val memzero: top -> uint32 -> unit

Where top stands for a top type. This then gets (conveniently)
translated to a declaration

void memzero (void *dst, uint32_t len);

which is the intended behavior, since the top type allows encoding
buffer polymorphism à la C.

This doesn't work, though, because F* models a number of elements, while
the C implementation wants a number of bytes. By the time the pointer
has been cast to `void *`, we have lost the static type information and
can no longer convert a number of elements to a number of bytes.

So, KreMLin adds special treatment:
- at call-sites of `memzero`, the length is multiplied by a `sizeof`
- when generating the relevant header, we switch to a uint64_t instead
  of a uint32_t (we can't switch to a size_t, unfortunately, because by
  the time this replacement happens all names are still properly scoped
  and size_t is not something that was in the initial program scope).