Filter the headers a second time.
Initially, we decline to generate headers that contain no declarations.

After the most recent changes, it may now be the case that a header
contains declarations, but that all of these declarations are skipped
because they're all covered by a static-header option *and* a library
option (the new behavior).

Therefore, we filter headers a second time, then proceed to update the
dependencies of subsequent headers to account for the now-gone files.