Warning 23: macro collision
The naming from F* to C is generally injective, except for C macros,
which do not benefit from collision avoidance (for predictability of the
generated C code).

As such, there was a risk of undetected macro collisions when two
definitions marked as `CMacro` only differ by their case. This now
errors out with Warning 23.