Fix compilation of Steel_SpinLock_lock
The Steel SpinLock type violates the value semantics of Low*. Its
low-level implementation, using pthread, relies on the address of the
value; therefore, `lock` is a type that cannot be passed by value (it
would create new locks). In order to maintain the "identity" of a lock
(the address of the value of type Steel_SpinLock_lock), we pass it by
reference, therefore guaranteeing that the value lives only in a single
place in memory.