"convert" ctypes-test to dune
The purpose of this commit is to show that ctypes stubs
generated by Karamel's `-ctypes` option can be compiled
and used with dune.

However, this is not the recommended way to compile a program
with ctypes stubs. Indeed, here, all dependencies are written
manually in dune files (the ctypes.depend file generated by
Karamel is not used.)

By contrast, dune itself can automatically generate ctypes stubs
etc., which might be a substitute to Karamel's `-ctypes` option,
but this is not the purpose of this commit.