krmllib: Makefile: fix ignore of non-extracted modules
The makefile was wrongfully creating these files via the touch command,
which meant to just update the timestamp. Use -c to prevent creation of
files with this rule.

Also, for the modules which we do not want to extract, use --extract to
leave them out of the dependency analysis, so we do not need to ignore
them with the -bundle flag (which would also now fail, given that they
do not exist).

I added the FStar.Stubs namespace to the list, as I will introduce that
in soon-to-come F* PR.