Github actions CI
This changes the CI to be completely cloud hosted and independent of the
runner machine. I think all features are retained and the .docker
directory can be blasted away.

Note that soon the buildmachine will stop building F* images, so the old
workflow will go through the 'standalone' path and be slower.