Rework C compiler detection
This patch makes karamel pick, by default, 'cc' from the path as the C
compiler. This command can be overriden with the -cc option (taking the
name of a program in the PATH, or a full path to an executable), or the
CC environment variable (idem).

Since karamel passes specific options to different compilers, there is
now logic to detect the flavor of compiler that karamel is talking to
(e.g. by looking at `cc --version`).

The flavor can be overridden with -ccflavor.

There is a special case when the compiler is exactly "msvc" (either via
-cc or CC): this is not taken to be a command nor path, and we instead
use the wrapper script exactly as before.