Allow homeless karamel
This makes karamel find its relevant directories and helper files
without a need for KRML_HOME. The variable is no longer read. Karamel
will now just look for the files relative to its executable location,
which is uniform across 1) local builds 2) OPAM install 3) binary
packages (added by this PR).

There is a bit of a dance to get a local build of krmllib before karamel
is installed. We pass some environment variables to prevent the
autodetection since karamel is not yet in a cone that contains these
files (it's simply in the dune build directory).

One wart: the top-level 'krml' symlink is still here, but it points to
_build/default/src/Karamel.exe after 'make minimal' and to out/bin/krml
after a full 'make'. The minimal one will fail to find its supporting
(if it has to) and crash.

Also
1) Removes AutoConfig
2) Moves ./runtime to ./krmllib/runtime (it was already there in the
   installs)

(After this patch, karamel builds, but the tests don't fully run yet.)