PrintC: nit, group operands
This improves the printing so that the operands can break into multiples
lines independendently of each other, and of the full expression. This
is still not perfect, we could accumulate operands and `flow` them
instead of this, but for a one liner it's a nice improvement.

Example, for:
```fstar
let foo (long_name : U64.t) : U64.t =
  let open FStar.UInt64 in
  (long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name)
  -%^ (long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name -%^ long_name)
```

The code goes from
```c
uint64_t foo(uint64_t long_name)
{
  return
    long_name
    - long_name
    - long_name
    - long_name
    - long_name
    - long_name
    - long_name
    - (long_name - long_name - long_name - long_name - long_name - long_name - long_name);
}
```
To:
```c
uint64_t foo(uint64_t long_name)
{
  return
    long_name - long_name - long_name - long_name - long_name - long_name - long_name -
      (long_name - long_name - long_name - long_name - long_name - long_name - long_name);
}
```