Fix subtlety linked to array decay rules in C.
Desugaring { ...; f = e; ... } into:

```
T x;
...
x.f = e;
...
```

One must make sure that in the internal AST, the type of the LHS of the
assignment is the type of `f` as defined in its type definition, not the
type of `e`.

Barring that, in the presence of array decay, this might result in an
assignment at array type, which might be misinterpreted to mean a copy,
as opposed to an assignment at pointer type, which relies on the RHS
decaying (which if fine).