Introduce -faggressive-inlining
This flag makes karamel attempt to inline every intermediate letbinding,
instead of only those marked with `@@CInline` or with some particular
name.

There is an F* flag that is similar (`--ext extraction_inline_all`) that
adds a `@@CInline` attribute to every letbinding, but using that flag
means the choice of whether to inline or not makes it into the .checked
file, and changing it requires re-typechecking. Further, it means we
cannot reuse a single krml file to obtained both inlined and non-inlined
code, for example C and Rust.