CStarToC11: fix purity check for VLA length
Sorry, I missed the fact that arithmetic expression are calls to
operators, so this patch makes sure we detect them as pure. Also, tweak
the definition of `mul` to constant-fold a size of 1, to avoid diffs
like this:

```diff
-  uint32_t *one = (uint32_t *)alloca(len * sizeof (uint32_t));
+  uint32_t *one = (uint32_t *)alloca(len * (1U * sizeof (uint32_t)));
```

With this I do not see a diff in HACL* (so there are zero non-pure array
lengths, which is a good sign).