CI: Use OCaml 5.3.0
There is currently a problem with OCaml's pprintast/ppxlib in that it
generates "raw identifiers" like `\#module` which are not valid in OCaml
below 5.2.0

See https://github.com/ocaml/opam-repository/pull/28610#issuecomment-3362301901 for some discussion.

I think this will be fixed upstream, but bumping to OCaml 5.3.0 seems
useful anyway. (We can also work around it by using `" module"` instead
of `"module"` in lib/GenCtypes.ml, but that's nasty.)