GenCtypes: fix generation of module expressions
The previous code was using an identifier called "module" to generate
a "module" keyword. Newer OCaml's will use a raw identifier here
("\#module") to avoid the clash with the keyword when an identifier
is really needed. But, we actually want the keyword here, so use the
appropriate constructor.

This is the failure I mentioned in the previous CI patch. I have no idea
why bumping to 5.3.0 made it work in that branch, but later break in
the scheduled build. I think this patch should also work fine in older
OCaml's.