Introduce a constant-folding / simplification phase
This phase will fold operations between constants that we not reduced by
F* (e.g. if they were enabled by inlining at the Karamel level) and also
simplify expressions such as `x+0` or `x*1` which may appear at the C
level.

The correctness of the constand folding relies on the absence of
overflow and UB in these operations (except for the wrapping operators,
these are handled well), and the fact the operations in C and OCaml's
ZArith coincide. It's my understanding that the semantics of division
and modulo in C and ZArith coincide
(https://ocaml.org/p/zarith/latest/doc/zarith/Z/index.html#val-div, C
standard s6.5.6).